requirements.txt
altair
numpy
pandas
pydeck
streamlit
langchain==0.0.348
numpy==1.26.2
python-dotenv==1.0.0